# optimized cross product