@kooparse 2025-05-11
Thanks, and it seems indeed more efficient, but there's also another cost to it; readability. This library wants to be very readable, and I would even sometime put more weight on that compared to other, most optimise, solutions.

This PR complexity too much the body of the cross product function in my opinion. Nonetheless, everyone is free to add his own custom implementation of this lib methods in their own codebase! 

Thanks again tho!