@Justus2308 2025-05-11
Yeah that's a completely valid concern, the code doesn't really 'look' like it's performing a cross product anymore.
Of course everyone is free to use this implementation in their own code anyways if they are ok with that :)